+++ R/plot_chromHMM.R
@@ -2,7 +2,9 @@
 #'
 #' Creates a heatmap using outputs from ChromHMM using ggplot2.The function
 #' takes a list of peakfiles, performs ChromHMM and outputs a heatmap. ChromHMM
-#' annotation file must be loaded prior to using this function.
+#' annotation file must be loaded prior to using this function. 
+#' ChromHMM annotations are aligned to hg19, and will be automatically lifted 
+#' over to the \code{genome_build} to match the build of the \code{peaklist}.
 #'
 #' @param peaklist A named \link[base]{list} of peak files as GRanges object. 
 #' If list is not named, default names will be assigned.
@@ -15,14 +17,14 @@
 #' peakfiles. "hg19" or "hg38".
 #' @param interact Default TRUE. By default, the heatmaps are interactive.
 #' If\code{FALSE}, the function generates a static ChromHMM heatmap.
+#' @param return_data Return the plot data as in addition to the plot itself.
+#' @return ChromHMM heatmap, or a named list. 
 #'
-#' @return ChromHMM heatmap
-#'
+#' @importMethodsFrom genomation annotateWithFeatures 
+#' @importFrom genomation heatTargetAnnotation 
 #' @importFrom GenomicRanges GRangesList
-#' @importMethodsFrom genomation annotateWithFeatures
-#' @importFrom genomation heatTargetAnnotation
 #' @importFrom reshape2 melt
-#' @importFrom plotly plot_ly
+#' @importFrom plotly ggplotly
 #' @import ggplot2
 #' @importMethodsFrom rtracklayer liftOver
 #' @importFrom AnnotationHub AnnotationHub
@@ -31,22 +33,21 @@
 #' @examples
 #' data("CnT_H3K27ac") # example dataset as GRanges object
 #' data("CnR_H3K27ac") # example dataset as GRanges object 
-#' peaks <- list(CnT=CnT_H3K27ac, CnR=CnR_H3K27ac) # create a list
+#' peaklist <- list(CnT=CnT_H3K27ac, CnR=CnR_H3K27ac) # create a list
 #' 
-#' my_plot <- plot_chromHMM(peaklist=peaks, 
+#' my_plot <- plot_chromHMM(peaklist = peaklist, 
 #'                          cell_line = "K562",
 #'                          genome_build = "hg19") 
 plot_chromHMM <- function(peaklist,
                           chromHMM_annotation,
                           cell_line=NULL,
                           genome_build,
-                          interact=TRUE){
+                          interact=TRUE,
+                          return_data=FALSE){
   # define variables
-  chain <- NULL
-  State <- NULL
-  Sample <- NULL
-  value <- NULL
+  State <- Sample <- value <- NULL;
    
+  message("--- Running plot_chromHMM() ---")
   #### Automatically get cell line data ####
   if(!is.null(cell_line) &&
      !is.null(check_cell_lines(cell_lines = cell_line,
@@ -57,39 +58,37 @@ plot_chromHMM <- function(peaklist,
       stp <- "Must supply chromHMM_annotation or cell_line."
       stop(stp)
   }
+  #### Liftover ####
+  ## chromHMM_annotation must be lifted over to match genome build of peaklist.
+  chromHMM_annotation <- liftover_grlist(grlist = chromHMM_annotation, 
+                                         input_build = "hg19", 
+                                         output_build = genome_build, 
+                                         as_grangeslist = TRUE)
+  
   # check that peaklist is named, if not, default names assigned
-  peaklist <- check_list_names(peaklist)
-  # check that there are no empty values
-  i <- 1
-  while(i <= length(peaklist)){
-    if (length(peaklist[[i]])==0){
-      peaklist[[i]] <- NULL
-    }else{
-      i <- i + 1
-    }
-  }
-  peaklist <- liftover_grlist(grlist = peaklist, 
-                              input_build = genome_build, 
-                              output_build = "hg19")
-  # create GRangeList from GRanges objects
-  grange_list <- GenomicRanges::GRangesList(peaklist, compress = FALSE)
+  peaklist <- prepare_peaklist(peaklist = peaklist, 
+                               remove_empty = TRUE, 
+                               as_grangeslist = TRUE) 
   # annotate peakfiles with chromHMM annotations
-  annotation <- genomation::annotateWithFeatures(grange_list,
-                                                 chromHMM_annotation)
+  message("Annotating with features.")
+  annotation <- genomation::annotateWithFeatures(
+      target = peaklist,
+      features = chromHMM_annotation)
   # obtain matrix
+  message("Obtaining target annotation matrix.")
   matrix <- genomation::heatTargetAnnotation(annotation, plot = FALSE)
-  rownames(matrix) <- names(peaklist) # set row names
   # remove numbers in front of states
   label_corrected <- gsub('X', '', colnames(matrix))
-  colnames(matrix) <- label_corrected # set corrected labels
-  # if interaction is FALSE generate static heatmap
-  if(!interact){
+  colnames(matrix) <- label_corrected # set corrected labels 
     # convert matrix into molten data frame
     matrix_melt <- reshape2::melt(matrix)
     colnames(matrix_melt) <- c("Sample", "State", "value")
     # create heatmap
     chrHMM_plot <- ggplot2::ggplot(matrix_melt) +
-      ggplot2::geom_tile(ggplot2::aes(x = State, y = Sample, fill = value)) +
+      ggplot2::geom_tile(
+          ggplot2::aes(x = State, 
+                       y = Sample, 
+                       fill = value)) +
       ggplot2::ylab("") +
       ggplot2::xlab("") +
       ggplot2::scale_fill_viridis_b() +
@@ -98,14 +97,20 @@ plot_chromHMM <- function(peaklist,
       ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 315,
                                                          vjust = 0.5,
                                                          hjust=0))
-  }else{
-    # generate interactive heatmap
-    chrHMM_plot <- plotly::plot_ly(x = colnames(matrix),
-                                   y = rownames(matrix),
-                                   z = matrix,
-                                   type = "heatmap")
-  }
-  # return heatmap
-  return(chrHMM_plot)
+    #### Make plot interactive ####
+    if(isTRUE(interact)){
+        chrHMM_plot <- plotly::ggplotly(chrHMM_plot)
+    } 
+  #### Return ####
+    if(return_data){
+        message("Returning named list with both plot and data.")
+        return(
+            list(plot=chrHMM_plot, 
+                 data=matrix_melt)
+        )
+    } else{
+        message("Returning chrHMM_plot.")
+        return(chrHMM_plot)
+    } 
 }